cu29_clock

Struct CuDuration

pub struct CuDuration(pub u64);

For Robot times, the underlying type is a u64 representing nanoseconds. It is always positive to simplify the reasoning on the user side.

Tuple Fields
0: u64
Implementations

impl CuDuration

pub const MIN: CuDuration

pub const MAX: CuDuration

pub fn max(self, other: CuDuration) -> CuDuration

pub fn min(self, other: CuDuration) -> CuDuration

pub fn as_nanos(&self) -> u64

pub fn as_micros(&self) -> u64

pub fn as_millis(&self) -> u64

pub fn as_secs(&self) -> u64

pub fn from_nanos(nanos: u64) -> Self

pub fn from_micros(micros: u64) -> Self

pub fn from_millis(millis: u64) -> Self

pub fn from_secs(secs: u64) -> Self