Crate cu29_clock

Structs

CuDuration

For Robot times, the underlying type is a u64 representing nanoseconds. It is always positive to simplify the reasoning on the user side.

CuTimeRange

Represents a time range.

Instant

A point-in-time wall-clock measurement.

OptionCuTime

Homebrewed Option<CuDuration> to avoid using 128bits just to represent an Option.

PartialCuTimeRange

Represents a time range with possible undefined start or end or both.

RobotClock

A running Robot clock. The clock is a monotonic clock that starts at an arbitrary reference time. It is clone resilient, ie a clone will be the same clock, even when mocked.

RobotClockMock

A mock clock that can be controlled by the user.

Enums

Tov

The time of validity of a message can be more than one time but can be a time range of Tovs. For example a sub scan for a lidar, a set of images etc… can have a range of validity.

Traits

ClockProvider

A trait to provide a clock to the runtime.

Type Aliases

CuTime

A robot time is just a duration from a fixed point in time.